pyproject.toml
[project]
name = "langgraph-project-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=1.1.6",
    "notebook>=7.5.5",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]
